fix: update to pl-1.3.0, fix type of droupout

+++ bttr/model/bttr.py
@@ -19,7 +19,7 @@ class BTTR(pl.LightningModule):
         nhead: int,
         num_decoder_layers: int,
         dim_feedforward: int,
-        dropout: int,
+        dropout: float,
     ):
         super().__init__()
